ww/requirements.txt
streamlit>=1.32.0
chromadb>=0.4.22
langchain>=0.1.9
langchain-community>=0.0.24
pypdf>=4.0.1
python-docx>=1.1.0
python-magic>=0.4.27
ollama>=0.1.6
python-dotenv>=1.0.1
loguru>=0.7.2 